Moved cox files and hide Base classes
