+++ pycox/models/__init__.py
@@ -1,7 +1,7 @@
 
-from pycox.models import base, loss, utils, pmf
-from pycox.models.cox.cox import CoxPH
-from pycox.models.cox.cox_cc import CoxCC, CoxTime
+from pycox.models import base, loss, utils, pmf, data
+from pycox.models.cox import CoxPH
+from pycox.models.cox_cc import CoxCC, CoxTime
 from pycox.models.deephit import DeepHitSingle, DeepHit
 from pycox.models.pmf import PMF
 from pycox.models.logistic_hazard import LogisticHazard
